chore: simplify namespace
